requirements.txt
streamlit==1.2.0
Pillow==10.0.0
pandas==1.3.5
numpy==1.19.5